requirements.txt
PyYAML==3.10
distribute==0.6.28
